+++ DaoGenerator/src/de/greenrobot/daogenerator/DaoUtil.java
@@ -104,5 +104,11 @@ public class DaoUtil {
         return byteCount;
     }
 
-
+    public static String checkConvertToJavaDoc(String javaDoc, String indent) {
+        if (javaDoc != null && !javaDoc.trim().startsWith("/**")) {
+            javaDoc = javaDoc.replace("\n", "\n" + indent + " * ");
+            javaDoc = indent + "/**\n" + indent + " * " + javaDoc + "\n" + indent + " */";
+        }
+        return javaDoc;
+    }
 }