+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release History
 * Deadlock prevention when loading a list of entities while doing concurrent updates
 * Fixed async queries
 * Better Android Studio support
+* Generator: Added defining JavaDoc for entities and their properties
 * Generator: Fixed codeBeforeGetter, added codeBeforeGetterAndSetter
 
 ### V2.0.0 Major feature release (2015-07-30, both core and generator)